Page::editableElementsInRect() should find nested editable elements and return found elements in paint order https://bugs.webkit.org/show_bug.cgi?id=208352 <rdar://problem/59867815>
Reviewed by Wenson Hsieh.

Source/WebCore:

Use the existing rect-based hit testing machinery to collect all the editable elements on
the page that a person can hit. This makes it possible to find nested editable elements
(e.g. an <input> inside a <div contenteditable="true">), ignore elements with CSS "pointer-events: none",
and return elements in paint order (closest to farthest to a person's face). The latter
makes it possible for a caller to know what element is frontmost, especially when two
editable elements overlap.

* page/Page.cpp:
(WebCore::Page::editableElementsInRect const): Implement in terms of hit testing machinery.
* rendering/HitTestLocation.cpp:
(WebCore::HitTestLocation::HitTestLocation):
* rendering/HitTestLocation.h:
* rendering/HitTestResult.cpp:
(WebCore::HitTestResult::HitTestResult):
* rendering/HitTestResult.h:
Added convenience constructors that take a LayoutRect.

Tools:

Update test results to reflect the new behavior. To do this I also need to fix up the test code
to actually scroll the web content. This also means we now execute the same code on both Mac and
iOS to do the scrolling: window.scrollTo(0, 5000). Also fixed up sub-test "Inputs scrolled outside
the requested rect; should not be included." to use a 10000px height <div> so that you can scroll
to a y position of 5000. Otherwise, it's impossible to do because the page height is < 5000px.

* TestWebKitAPI/Tests/WebKitCocoa/RequestTextInputContext.mm:
(TEST):

Identifier: 221381@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@257722 268f45cc-cd09-0410-ab3c-d52691b4dbfc