[WebDriver] Update screenshot/user_prompt expectations.
They were marked as FAIL in r239824 due to upstream w3c tests not handling
dialogs correctly and this handling not being in the spec.

The correct handling was added to wpt in
https://github.com/web-platform-tests/wpt/pull/17563, imported to webkit
in r252358, and have been passing since in the GTK test bot.

Also were added to the spec in https://github.com/w3c/webdriver/commit/32165846

Unreviewed test gardening.

* TestExpectations.json:

Identifier: 221526@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@257894 268f45cc-cd09-0410-ab3c-d52691b4dbfc