Page-specific UserStyleSheets should wait until the initial empty document has been removed to be injected https://bugs.webkit.org/show_bug.cgi?id=208644 <rdar://problem/60042429>
Patch by Antoine Quint <graouts@apple.com> on 2020-03-05
Reviewed by Brady Eidson.

Source/WebCore:

When a WKWebView is created and asked to load a URL, it will first load an empty initial document (about:blank) prior
to creating a Document for the requested URL. We need to ensure that page-specific UserStyleSheets are only injected
once the Document for the requested URL starts loading.

When Page::injectUserStyleSheet() is called, if we determine that the empty initial document is visible, we enqueue
the provided UserStyleSheet and wait until the new mainFrameDidChangeToNonInitialEmptyDocument() method is called
to empty that queue and inject the UserStyleSheets then.

This new method is called from Frame::setDocument() for a main frame once we've determined the frame's document has
changed to a non-null value and that the frame loader's state machine indicates that we're no longer displaying the
initial empty document.

* dom/ExtensionStyleSheets.h:
* page/Frame.cpp:
(WebCore::Frame::setDocument):
* page/Page.cpp:
(WebCore::Page::injectUserStyleSheet):
(WebCore::Page::removeInjectedUserStyleSheet):
(WebCore::Page::mainFrameDidChangeToNonInitialEmptyDocument):
* page/Page.h:

Tools:

Add new tests for -[_WKUserStyleSheet initWithSource:forWKWebView:forMainFrameOnly:userContentWorld:]) that check:

1. that a _WKUserStyleSheet can be added immediately after a WKWebView's creation and will be injected once the initial
   empty document has been removed in favor of the document for the requested URL,
2. that removing a _WKUserStyleSheet immediately after it was added but before the initial empty document was removed
   correctly does not injected the style sheet.

* TestWebKitAPI/Tests/WebKitCocoa/UserContentController.mm:

Identifier: 221570@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@257939 268f45cc-cd09-0410-ab3c-d52691b4dbfc