Add signposts for top-level execution of script elements https://bugs.webkit.org/show_bug.cgi?id=208548
Reviewed by Alex Christensen.

This adds signposts for initial global scope code evaluation of script elements. This makes
it easier to understand what code is executing and blocking initial HTML parsing from
completing.

In the long run, it would be nice to extend this to all top-level entry points back into JS
(e.g. we should emit an interval if we re-enter a script via setTimeout). However, such
probes have been removed in the past for being too noisy or slow (like in
https://bugs.webkit.org/show_bug.cgi?id=187196), so let's just start with this for now.

Note that we do emit the script URL in the signpost, but the signpost is gated behind both
an environment variable and an Apple-internal check to prevent accidentally leaking
sensitive info.

* dom/ScriptElement.cpp:
(WebCore::ScriptElement::executeClassicScript):
(WebCore::ScriptElement::executeModuleScript):


Identifier: 221586@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@257955 268f45cc-cd09-0410-ab3c-d52691b4dbfc