FocusController::setFocusedElement() should tell client of refocused element https://bugs.webkit.org/show_bug.cgi?id=208880
Reviewed by Wenson Hsieh.

Source/WebCore:

If the specified new focus element is non-nullptr and is already focused then tell the client
that the element was re-focused so that it may update its input state, if needed. On iOS, this
lets the UI process evaluate again whether to start an input session (i.e. bring up the keyboard),
which may have been disallowed when the element was originally focused (say, it was programmatically
focused and there was no hardware keyboard attached).

* page/FocusController.cpp:
(WebCore::FocusController::setFocusedElement):

Tools:

Adds a test to ensure that calling -selectPositionAtPoint on an already focused element tries
to start an input session. Currently, it does not even try.

* TestWebKitAPI/Tests/ios/UIWKInteractionViewProtocol.mm:
(TEST):

Identifier: 221901@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@258333 268f45cc-cd09-0410-ab3c-d52691b4dbfc