AXIsolatedTree removal should set all nodes to be removed on AX secondary thread. https://bugs.webkit.org/show_bug.cgi?id=209169
Reviewed by Chris Fleizach.

- AXIsolatedTree::removeTreeForPageID is called on the main thread but
it should not remove the nodes in the main thread, but instead add them
to the pending changes to be removed on the secondary thread. This was
causing the problem of empty new trees when the old tree would go away
but the client was holding a reference to an object that has been
disconnected and thus had no children.
- In addition, this change fixes an isolated tree mode crash in AccessibilityMenuList.

* accessibility/AXObjectCache.cpp:
(WebCore::AXObjectCache::~AXObjectCache):
* accessibility/AccessibilityMenuList.cpp:
(WebCore::AccessibilityMenuList::isCollapsed const):
* accessibility/isolatedtree/AXIsolatedTree.cpp:
(WebCore::AXIsolatedTree::removeTreeForPageID):


Identifier: 222098@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@258560 268f45cc-cd09-0410-ab3c-d52691b4dbfc